+++ src/commands/rotate/infra_cert.rs
@@ -214,6 +214,8 @@ mod tests {
             openbao_advertise_addr: None,
             http01_admin_bind_addr: None,
             http01_admin_advertise_addr: None,
+            stepca_bind_addr: None,
+            stepca_advertise_addr: None,
             infra_certs,
         };
 
@@ -340,6 +342,8 @@ mod tests {
             openbao_advertise_addr: None,
             http01_admin_bind_addr: Some("192.168.1.10:8080".to_string()),
             http01_admin_advertise_addr: None,
+            stepca_bind_addr: None,
+            stepca_advertise_addr: None,
             infra_certs,
         };